mostly done

+++ RESTAPIs/controllers/feed.js
@@ -4,6 +4,7 @@ const fs = require("fs");
 const path = require("path");
 
 const Post = require("../models/post.js");
+const User = require("../models/user.js");
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -46,18 +47,28 @@ exports.createPosts = (req, res, next) => {
   const imageUrl = req.file.path.replace("\\", "/");
   const title = req.body.title;
   const content = req.body.content;
+  let creator;
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: { name: "Amogh" },
+    creator: req.userId,
   });
   post
     .save()
     .then((result) => {
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then((result) => {
       res.status(201).json({
         message: "Post created successfully",
-        post: result,
+        post: post,
+        creator: { _id: creator._id, name: creator.name },
       });
     })
     .catch((err) => {
@@ -107,7 +118,12 @@ exports.updatePost = async (req, res, next) => {
     const post = await Post.findById(postId);
     if (!post) {
       const error = new Error("Post not found");
-      error.statusCode = 422;
+      error.statusCode = 402;
+      throw err;
+    }
+    if (post.creator.toString() !== req.userId) {
+      const error = new Error("Not authorized User!");
+      error.statusCode = 403;
       throw err;
     }
     if (!imageUrl) {
@@ -142,12 +158,23 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 422;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized User!");
+        error.statusCode = 403;
+        throw error;
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndDelete(postId);
     })
-    .then((result) => {
-      console.log(result);
-      res.status(200).json({ message: "Post deleted successfully" });
+    .then(result => {
+      return User.findById(req.userId);
+    })
+    .then(user =>{
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(result => {
+      res.status(200).json({message: 'Post Deleted Successfully!'})
     })
     .catch((err) => {
       if (!err.statusCode) {
